CHAM-64 FEAT: 챗봇 페이지 컴포넌트 분리
