+++ components/chat/ChatHeader.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const ChatHeader = () => {
+  return (
+    <div>
+
+    </div>
+  );
+};
+
+export default ChatHeader;
